backend/requirements.txt
# Core FLUX.1-dev dependencies - pinned versions for compatibility
torch>=2.0.0
torchvision>=0.15.0
diffusers==0.32.1
transformers==4.47.1
accelerate==1.2.1
sentencepiece==0.2.0
protobuf==5.29.2

# ControlNet and advanced conditioning
controlnet-aux>=0.0.7
opencv-python>=4.8.0

# Web server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Image processing dependencies
Pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0

# HTTP client for testing
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Optional: Better logging
colorlog>=6.7.0

# Optional: Performance monitoring
psutil>=5.9.0
